Release/1.5.0

+++ src/datarecord.cpp
@@ -592,10 +592,10 @@ void datarecord::schedule(reclist& thisi, double maxtime,
  * infusion end. 
  * 
  */
-void insert_record(reclist& thisi, const int start, rec_ptr& rec, 
+void insert_record(reclist& thisi, const size_t start, rec_ptr& rec, 
                    const bool put_ev_first) {
   double time = rec->time();
-  int i = start;
+  size_t i = start;
   if(put_ev_first) {
     for(i = start + 1; i < thisi.size(); ++i) {
       if(thisi[i]->time() >= time) {
